fix: change type of sum
- change type of sum from int to long for bigger number

+++ src/main/java/calculator/Calculator.java
@@ -3,8 +3,8 @@ package calculator;
 import java.util.ArrayList;
 
 public class Calculator {
-    public static int sum(ArrayList<Integer> numbers) {
-        int sum = 0;
+    public static long sum(ArrayList<Integer> numbers) {
+        long sum = 0;
         for (Integer number : numbers) {
             sum += number;
         }
